requirements.txt
flask>=2.3.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
pywebview>=4.4.0
